+++ src/hysim/util/logging.py
@@ -1,0 +1,30 @@
+from contextlib import contextmanager
+
+import mitsuba as mi
+
+
+class CustomMitsubaFormatter(mi.Formatter):
+    def __init__(self, frame_index: int):
+        super().__init__()
+        self.frame_index = frame_index
+
+    def format(self, level: mi.LogLevel, thread, class_, file, line, msg):
+        return f" {level.name.upper():8} {chr(0x02523)}{chr(0x02501)} Frame {self.frame_index} - {msg}"
+
+    @staticmethod
+    @contextmanager
+    def log(frame_index: int):
+        """Creates a custom mitsuba formatter to match HySim and sets mitsuba's logger to use
+        it. Also sets the log level to mi.LogLevel.Info to make sure the start and finished
+        rendering messages are displayed."""
+
+        # NOTE: There is no progress bar displayed in the console if mi.variant() is not a scalar variant
+        mitsuba_logger = mi.Thread.thread().logger()
+        log_level = mitsuba_logger.log_level()
+        try:
+            mitsuba_logger.set_formatter(CustomMitsubaFormatter(frame_index))
+            mitsuba_logger.set_log_level(mi.LogLevel.Info)
+            yield
+        finally:
+            mitsuba_logger.set_log_level(log_level)
+            del mitsuba_logger
